backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
firebase-admin==6.2.0
pymongo==4.11.0
motor==3.7.0
python-dotenv==1.0.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==1.10.12
bcrypt==4.0.1
tensorflow>=2.0.0
keras==2.15.0
numpy
imagehash==4.2.1
Pillow==10.1.0
requests==2.31.0
faiss-cpu==1.9.0
scikit-image

